Added spectrogram
It is in a very poor state, but it exists. Next commits will be focusing on improving the quality of the spectrogram. Namely, making it vertical instead of horizontal and also increasing the audio transmission rate specifically for the spectrogram but not for the audio analysis, that needs to be in longer cycles.